Added 3rd Person Controller Package
Downloaded Vinicius Marques' 3rd person controller including fly mode.  Needs to be incorporated into the test scene and adjusted for the game.  Needs a scroll wheel zoom in and out, at minimum.  May want to remove aim-mode, althought that may give a more seamless transition from ranged to melee when and enemy is targeted and could be a fun control.  Base speed should be the left click speed, and left click should be reserved for world interactions.