Incorporated New Player Controller
Downloaded and added the player controller to the test scene.  Commented out the slow walk and the aim functionality so that the left click, right click and middle click can be better used.  World interactions needs an update for the new input manager (since mouse clicks have been reassigned).  Zoom functionality needs to be added to the camera script.